+++ src/server/graphql/character/index.js
@@ -1,9 +1,15 @@
 import characters from './characters.query';
+import toggleCharacter from './toggleCharacter.mutation';
 
 const queries = {
   characters,
 };
 
+const mutations = {
+  toggleCharacter,
+};
+
 export default {
   queries,
+  mutations,
 };
