+++ src/server/graphql/character/toggleCharacter.mutation.js
@@ -1,10 +1,6 @@
-/**
- * Created by mikhail on 25.11.15.
- */
 import {GraphQLNonNull, GraphQLString, GraphQLBoolean} from 'graphql';
 import {mutationWithClientMutationId, fromGlobalId} from 'graphql-relay';
-import characters from '../queries/queries/characters';
-import {toggleCharacter} from '../../../data/characters.js';
+import viewer from '../viewer';
 
 export default mutationWithClientMutationId({
   name: 'ToggleCharacter',
@@ -18,14 +14,12 @@ export default mutationWithClientMutationId({
       description: 'Character id'
     }
   },
-  outputFields: {
-    characters
-  },
-  mutateAndGetPayload({characterId, kill}) {
-    let {id} = fromGlobalId(characterId);
-
-    // revive in list
-    toggleCharacter(id, kill);
+  outputFields: () => ({ viewer }),
+  mutateAndGetPayload() {
+    // let {id} = fromGlobalId(characterId);
+    //
+    // // revive in list
+    // toggleCharacter(id, kill);
 
     return {};
   }
