+++ src/server/graphql/mutations.js
@@ -1,0 +1,5 @@
+import character from './character';
+
+export default {
+  ...character.mutations,
+}
